refactor: rename selectedSlots.ts to selectedSlot.ts and update all imports
