+++ apps/api/v2/src/lib/services/available-slots.service.ts
@@ -3,11 +3,12 @@ import { PrismaEventTypeRepository } from "@/lib/repositories/prisma-event-type.
 import { PrismaOOORepository } from "@/lib/repositories/prisma-ooo.repository";
 import { PrismaRoutingFormResponseRepository } from "@/lib/repositories/prisma-routing-form-response.repository";
 import { PrismaScheduleRepository } from "@/lib/repositories/prisma-schedule.repository";
+import { PrismaSelectedSlotRepository } from "@/lib/repositories/prisma-selected-slots.repository";
 import { PrismaTeamRepository } from "@/lib/repositories/prisma-team.repository";
 import { PrismaUserRepository } from "@/lib/repositories/prisma-user.repository";
 import { Injectable } from "@nestjs/common";
 
-import { SelectedSlotRepositoryInterface } from "@calcom/lib/server/repository/selectedSlots";
+import { SelectedSlotRepositoryInterface } from "@calcom/lib/server/repository/selectedSlot";
 import { AvailableSlotsService as BaseAvailableSlotsService } from "@calcom/platform-libraries/slots";
 
 @Injectable()
@@ -18,7 +19,7 @@ export class AvailableSlotsService extends BaseAvailableSlotsService {
     private readonly teamRepository: PrismaTeamRepository,
     private readonly routingFormResponseRepository: PrismaRoutingFormResponseRepository,
     private readonly bookingRepository: PrismaBookingRepository,
-    private readonly selectedSlotRepository: SelectedSlotRepositoryInterface,
+    private readonly selectedSlotRepository: PrismaSelectedSlotRepository,
     private readonly eventTypeRepository: PrismaEventTypeRepository,
     private readonly userRepository: PrismaUserRepository
   ) {
