+++ packages/lib/di/modules/selectedSlots.ts
@@ -1,7 +1,7 @@
 import { createModule } from "@evyweb/ioctopus";
 
 import { DI_TOKENS } from "@calcom/lib/di/tokens";
-import { PrismaSelectedSlotRepository } from "@calcom/lib/server/repository/selectedSlots";
+import { PrismaSelectedSlotRepository } from "@calcom/lib/server/repository/selectedSlot";
 
 export const selectedSlotsRepositoryModule = createModule();
 selectedSlotsRepositoryModule
