+++ packages/platform/libraries/repositories.ts
@@ -3,6 +3,6 @@ export { ScheduleRepository as PrismaScheduleRepository } from "@calcom/lib/serv
 export { BookingRepository as PrismaBookingRepository } from "@calcom/lib/server/repository/booking";
 export { EventTypeRepository as PrismaEventTypeRepository } from "@calcom/lib/server/repository/eventTypeRepository";
 export { RoutingFormResponseRepository as PrismaRoutingFormResponseRepository } from "@calcom/lib/server/repository/formResponse";
-export { PrismaSelectedSlotRepository as PrismaSelectedSlotsRepository } from "@calcom/lib/server/repository/selectedSlots";
+export { PrismaSelectedSlotRepository as PrismaSelectedSlotsRepository } from "@calcom/lib/server/repository/selectedSlot";
 export { TeamRepository as PrismaTeamRepository } from "@calcom/lib/server/repository/team";
 export { UserRepository as PrismaUserRepository } from "@calcom/lib/server/repository/user";
