+++ packages/trpc/server/routers/viewer/slots/reserveSlot.handler.ts
@@ -5,7 +5,7 @@ import { v4 as uuid } from "uuid";
 import dayjs from "@calcom/dayjs";
 import { WEBAPP_URL } from "@calcom/lib/constants";
 import { MINUTES_TO_BOOK } from "@calcom/lib/constants";
-import { PrismaSelectedSlotRepository } from "@calcom/lib/server/repository/selectedSlots";
+import { PrismaSelectedSlotRepository } from "@calcom/lib/server/repository/selectedSlot";
 import type { PrismaClient } from "@calcom/prisma";
 import { BookingStatus } from "@calcom/prisma/enums";
